financial_dashboard_app_pro-3/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.4
MarkupSafe==2.1.5
pandas==2.2.2
APScheduler==3.10.4

pdfplumber==0.11.4
ics==0.7.2
